+++ api/src/data_migration/transformation/subtask/transform_opportunity.py
@@ -1,0 +1,81 @@
+import logging
+from typing import Tuple
+
+import src.data_migration.transformation.transform_constants as transform_constants
+import src.data_migration.transformation.transform_util as transform_util
+from src.data_migration.transformation.subtask.abstract_transform_subtask import (
+    AbstractTransformSubTask,
+)
+from src.db.models.opportunity_models import Opportunity
+from src.db.models.staging.opportunity import Topportunity
+
+logger = logging.getLogger(__name__)
+
+
+class TransformOpportunity(AbstractTransformSubTask):
+    def transform_records(self) -> None:
+        # Fetch all opportunities that were modified
+        # Alongside that, grab the existing opportunity record
+        opportunities: list[Tuple[Topportunity, Opportunity | None]] = self.fetch(
+            Topportunity,
+            Opportunity,
+            [Topportunity.opportunity_id == Opportunity.opportunity_id],
+        )
+
+        for source_opportunity, target_opportunity in opportunities:
+            try:
+                self.process_opportunity(source_opportunity, target_opportunity)
+            except ValueError:
+                self.increment(
+                    transform_constants.Metrics.TOTAL_ERROR_COUNT,
+                    prefix=transform_constants.OPPORTUNITY,
+                )
+                logger.exception(
+                    "Failed to process opportunity",
+                    extra={"opportunity_id": source_opportunity.opportunity_id},
+                )
+
+    def process_opportunity(
+        self, source_opportunity: Topportunity, target_opportunity: Opportunity | None
+    ) -> None:
+        self.increment(
+            transform_constants.Metrics.TOTAL_RECORDS_PROCESSED,
+            prefix=transform_constants.OPPORTUNITY,
+        )
+        extra = {"opportunity_id": source_opportunity.opportunity_id}
+        logger.info("Processing opportunity", extra=extra)
+
+        if source_opportunity.is_deleted:
+            self._handle_delete(
+                source_opportunity,
+                target_opportunity,
+                transform_constants.OPPORTUNITY,
+                extra,
+                error_on_missing_target=True,
+            )
+
+        else:
+            # To avoid incrementing metrics for records we fail to transform, record
+            # here whether it's an insert/update and we'll increment after transforming
+            is_insert = target_opportunity is None
+
+            logger.info("Transforming and upserting opportunity", extra=extra)
+            transformed_opportunity = transform_util.transform_opportunity(
+                source_opportunity, target_opportunity
+            )
+
+            if is_insert:
+                self.increment(
+                    transform_constants.Metrics.TOTAL_RECORDS_INSERTED,
+                    prefix=transform_constants.OPPORTUNITY,
+                )
+                self.db_session.add(transformed_opportunity)
+            else:
+                self.increment(
+                    transform_constants.Metrics.TOTAL_RECORDS_UPDATED,
+                    prefix=transform_constants.OPPORTUNITY,
+                )
+                self.db_session.merge(transformed_opportunity)
+
+        logger.info("Processed opportunity", extra=extra)
+        source_opportunity.transformed_at = self.transform_time
